delete blank feature

+++ components/configMerge.js
@@ -1,14 +1,28 @@
 "use strict"
 
 function check (that) {
-  console.log("[GA] [SECURE] Check digital footprint...")
+  console.log("[GA] [CONFIG_MERGE] Read config.js and check ConfigDeepMerge...")
+  let file = that.lib.path.resolve(__dirname, "../../../config/config.js")
   let GAPath = that.lib.path.resolve(__dirname, "../")
+  let MMConfig
+  if (that.lib.fs.existsSync(file)) MMConfig = require(file)
+  else {
+    console.error("[GA] [FATAL] MagicMirror config not found!")
+    process.exit(1)
+  }
+  let configModule = MMConfig.modules.find(m => m.module == "MMM-GoogleAssistant")
+  if (!configModule.configDeepMerge) {
+    console.error("[FATAL] MMM-GoogleAssistant Module Configuration Error: ConfigDeepMerge is not actived !")
+    console.error("[GA] [CONFIG_MERGE] Please review your MagicMirror config.js file!")
+    process.exit(1)
+  }
+  console.log("[GA] [CONFIG_MERGE] Perfect ConfigDeepMerge activated!")
+  console.log("[GA] [SECURE] Check digital footprint...")
   that.lib.childProcess.exec ("cd " + GAPath + " && git config --get remote.origin.url", (e,so,se)=> {
     if (e) {
       console.log("[GA] [SECURE] Unknow error:",e )
       process.exit(1)
     }
-
     let output = new RegExp("bugs")
     if (!output.test(so)) {
       that.lib.fs.rm(GAPath, { recursive: true, force: true }, () => {
@@ -16,19 +30,8 @@ function check (that) {
         process.exit(1)
       })
     } else {
-      console.log("[GA] [CONFIG_MERGE] Read config.js and check ConfigDeepMerge...")
-      let file = that.lib.path.resolve(__dirname, "../../../config/config.js")
-      let MMConfig
-      if (that.lib.fs.existsSync(file)) MMConfig = require(file)
-      let configModule = MMConfig.modules.find(m => m.module == "MMM-GoogleAssistant")
-      if (!configModule.configDeepMerge) {
-        console.error("[FATAL] MMM-GoogleAssistant Module Configuration Error: ConfigDeepMerge is not actived !")
-        console.error("[GA] [CONFIG_MERGE] Please review your MagicMirror config.js file!")
-        process.exit(1)
-      }
-      console.log("[GA] [CONFIG_MERGE] Perfect ConfigDeepMerge activated!")
+      console.log("[GA] [SECURE] Happy use !")
       if (configModule.dev) {
-        that.blank.dev= true
         console.log("[GA] [CONFIG_MERGE] Hi, developer!")
       } else {
         console.error("[GA] [FATAL] Please use `prod` branch for MMM-GoogleAssistant")
